@@ -18,46 +18,46 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.3.1",
-    "@angular/common": "^4.3.1",
-    "@angular/compiler": "^4.3.1",
-    "@angular/core": "^4.3.1",
-    "@angular/forms": "^4.3.1",
-    "@angular/http": "^4.3.1",
-    "@angular/platform-browser": "^4.3.1",
-    "@angular/platform-browser-dynamic": "^4.3.1",
-    "@angular/router": "^4.3.1",
-    "@angular/material": "^2.0.0-beta.8",
-    "@angular/cdk": "^2.0.0-beta.8",
-    "@covalent/core": "1.0.0-beta.6",
-    "angular-2-local-storage": "^1.0.1",
-    "angular2-fontawesome": "^0.9.0",
-    "core-js": "^2.4.1",
+    "@angular/animations": "^5.1.2",
+    "@angular/cdk": "^5.0.2",
+    "@angular/common": "^5.1.2",
+    "@angular/compiler": "^5.1.2",
+    "@angular/core": "^5.1.2",
+    "@angular/forms": "^5.1.2",
+    "@angular/http": "^5.1.2",
+    "@angular/material": "^5.0.2",
+    "@angular/platform-browser": "^5.1.2",
+    "@angular/platform-browser-dynamic": "^5.1.2",
+    "@angular/router": "^5.1.2",
+    "@covalent/core": "1.0.0-rc.1",
+    "angular-async-local-storage": "^3.1.0",
+    "angular-font-awesome": "^3.1.1",
+    "core-js": "^2.5.3",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
-    "ng2-ace-editor": "0.2.3",
-    "rxjs": "^5.1.0",
+    "ng2-ace-editor": "0.3.4",
+    "rxjs": "^5.5.6",
     "spinkit": "^1.2.5",
-    "zone.js": "^0.8.4"
+    "zone.js": "^0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "1.2.3",
-    "@angular/compiler-cli": "^4.3.1",
-    "@types/jasmine": "2.5.53",
-    "@types/node": "8.0.15",
-    "codelyzer": "3.1.2",
+    "@angular/cli": "^1.6.2",
+    "@angular/compiler-cli": "^5.1.2",
+    "@types/jasmine": "2.8.2",
+    "@types/node": "8.5.2",
+    "codelyzer": "4.0.2",
     "http-proxy-middleware": "^0.17.4",
-    "jasmine-core": "2.6.4",
-    "jasmine-spec-reporter": "4.1.1",
-    "karma": "1.7.0",
+    "jasmine-core": "2.8.0",
+    "jasmine-spec-reporter": "4.2.1",
+    "karma": "2.0.0",
     "karma-chrome-launcher": "2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "1.3.0",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "1.3.3",
+    "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.0",
-    "ts-node": "3.2.1",
-    "tslint": "5.5.0",
-    "typescript": "^2.4.2"
+    "protractor": "~5.2.2",
+    "ts-node": "4.1.0",
+    "tslint": "5.8.0",
+    "typescript": "2.5.3"
   }
 }
