@@ -17,8 +17,7 @@
 
 import {Injectable} from "@angular/core";
 import {ActivatedRouteSnapshot, Resolve, Router, RouterStateSnapshot} from "@angular/router";
-import {Observable} from "rxjs/Observable";
-import "rxjs/add/operator/toPromise";
+import {Observable} from "rxjs/Rx";
 import {ZNodeData} from "./znode-data";
 import {ZNodeMetaWith} from "../meta/znode-meta-with";
 import {ZNodeService} from "../../znode.service";
@@ -43,21 +42,11 @@ export class ZNodeDataResolver implements Resolve<ZNodeMetaWith<ZNodeData>> {
 
     return this.zNodeService
       .getData(nodePath)
-      .toPromise()
-      .catch((error) => {
-        this.feedbackService
-          .showError(error, null)
-          .afterClosed()
-          .subscribe(
-            () => this.router.navigate(["/editor"])
-          );
-      })
-      .then((data) => {
-        if (!data) {
-          return Promise.reject("Couldn't fetch znode data");
-        }
-
-        return Promise.resolve(data);
-      });
+      .catch(err => this.feedbackService.showErrorAndThrowOnClose(err))
+      .catch(err =>
+        Observable
+          .fromPromise(this.router.navigate(["/editor"]))
+          .switchMap(() => Observable.throw(err))
+      );
   }
 }
