@@ -17,7 +17,7 @@
 
 import {Injectable} from "@angular/core";
 import {RouterStateSnapshot, ActivatedRouteSnapshot, CanDeactivate} from "@angular/router";
-import {Observable} from "rxjs/Observable";
+import {Observable} from "rxjs/Rx";
 import {CanDeactivateComponent} from "./can-deactivate-component";
 
 @Injectable()
