@@ -37,8 +37,8 @@ const editorRoutes = [
   {
     path: "editor",
     component: EditorComponent,
-    canActivate: [EditorGuard],
-    canActivateChild: [EditorGuard],
+    // canActivate: [EditorGuard],
+    // canActivateChild: [EditorGuard],
     resolve: {
       children: ZNodeChildrenResolver
     },
@@ -46,22 +46,22 @@ const editorRoutes = [
       {
         path: "node",
         component: ZNodeContainerComponent,
-        canActivate: [ZNodeContainerGuard],
-        canActivateChild: [ZNodeContainerGuard],
+        // canActivate: [ZNodeContainerGuard],
+        // canActivateChild: [ZNodeContainerGuard],
         children: [
           {
             path: "data",
             component: ZNodeDataComponent,
-            canActivate: [ZNodeDataGuard],
-            canDeactivate: [CanDeactivateComponentGuard],
-            resolve: {
-              data: ZNodeDataResolver
-            }
+            // canActivate: [ZNodeDataGuard],
+            // canDeactivate: [CanDeactivateComponentGuard],
+            // resolve: {
+            //   data: ZNodeDataResolver
+            // }
           },
           {
             path: "acl",
             component: ZNodeAclComponent,
-            canDeactivate: [CanDeactivateComponentGuard],
+            // canDeactivate: [CanDeactivateComponentGuard],
             resolve: {
               acl: ZNodeAclResolver
             }
@@ -73,10 +73,10 @@ const editorRoutes = [
               meta: ZNodeMetaResolver
             }
           },
-          {
-            path: "**",
-            redirectTo: "data"
-          }
+          // {
+          //   path: "**",
+          //   redirectTo: "data"
+          // }
         ]
       }
     ]
