@@ -16,7 +16,7 @@
  */
 
 import {NgModule} from "@angular/core";
-import {MdIconModule, MdInputModule} from "@angular/material";
+import {MatIconModule, MatInputModule} from "@angular/material";
 import {CovalentCommonModule} from "@covalent/core";
 import {RegexpFilterComponent} from "./regexp/regexp-filter.component";
 import {WindowRef} from "./window/window-ref";
@@ -25,8 +25,8 @@ import {CanDeactivateComponentGuard} from "./guards";
 @NgModule({
   imports: [
     CovalentCommonModule,
-    MdIconModule,
-    MdInputModule
+    MatIconModule,
+    MatInputModule
   ],
   declarations: [
     RegexpFilterComponent
